+++ doc/CHANGELOG.rst
@@ -5,6 +5,8 @@ Changelog
 ------------------
 - Extended supported python and pysam versions
 - Adjusted to avoid performance problem when processing extremely deep pileups
+- Adjusted so that when no families pass filters show warning instead of
+  error message
 
 0.5 (9/13/2016)
 ---------------